requirements.txt
torch
torchvision
timm
opencv-contrib-python
comet_ml
pandas
numpy
Pillow
sklearn
matplotlib
dvc
easydict